// hostings/app/interfaces/SystemClasses/GoogleTypes.ts
enum Google {
  CALENDAR = "CALENDER.CALENDAR",
  CALENDAR_EVENT = "CALENDAR.EVENT",
  CALENDAR_EVENT_GUEST = "CALENDAR.EVENT_GUEST",
  GMAIL_DRAFT = "GMAIL.DRAFT",
  GMAIL_THREAD = "GMAIL.THREAD",
  GMAIL_MESSAGE = "GMAIL.MESSAGE",
  GMAIL_LABEL = "GMAIL.LABEL",
  GMAIL_ATTACHMENT = "GMAIL.ATTACHMENT",
  SPREADSHEET = "SPREADSHEET",
  SPREADSHEET_SHEET = "SPREADSHEET_SHEET",
  SPREADSHEET_RANGE = "SPREADSHEET_RANGE",
  BASE_BLOB = "BASE.BLOB",
  BASE_BLOBSOURCE = "BASE.BLOBSOURCE",
}

export default Google;